Don't store entity properties as we parse
As it is, the code responsible for this was quite inefficient, in
particular since it did an insane amount of allocations and copies. This
change decreased parsing time by over 90% for my test demo. It could have
been made faster, but after some thinking I've come to the conclusion
that this actually isn't a very useful thing to track. Since we don't
provide mid-parse hooks, all this can tell you is the entity state at
the very end of the demo, which isn't useful. Instead, we'll provide a
way to later track the entity state through to a certain tick. This can
be done efficiently, since we now construct a linked list of updates, so
can simply look at the latest delta and track backwards until we get a
full update. That could also be further optimised by allowing the user
to limit which entities are tracked, but this is for future mlugg to
think about.